requirements.txt
streamlit==1.28.0
openai==1.3.0
yt-dlp==2023.12.30
pydub==0.25.1
python-dotenv==1.0.0
requests==2.31.0
numpy>=1.19.3,<2.0.0